% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{louisiana}
\alias{louisiana}
\title{louisiana}
\format{
A tibble with eleven variables: \code{sample_number}, \code{depth_m},
\code{weight_g}, \code{volume_ml}, \code{lead_ppm}, \code{zinc_ppm}, \code{cadmium_ppm},
\code{chromium_ppm}, \code{nickel_ppm}, \code{iron_percent}, \code{aluminum_percent}
}
\usage{
louisiana
}
\description{
For more see \href{https://link.springer.com/article/10.1007/BF00398785}{Statistical Comparison of Heavy Metal Concentrations in Various Louisiana Sediments}
}
\keyword{datasets}
